Adds support for oraclesql language
When the Oracle Developer Tools for VS Code (SQL and PLSQL)(https://marketplace.visualstudio.com/items?itemName=Oracle.oracledevtools) extension is installed, *.sql files are associated with the oraclesql language.

The changes on this commit enable SQL Formatter to be used with the oraclesql language.

Based on https://github.com/kufii/vscode-sql-formatter/pull/36